Sorting, BFS
This was very challenging problem. I had very much fun solving it! 
Algo was to divide the points based on their orientation and sort the east going ones on the increasing order of y axis and north going ones on the increasing order of x-axis. Then to watch for collision and check if they're already stopped or not by the done array. And adding them to a graph to find the number of stopped cow because the number of stopped cow is transitive. I used bfs to find the number of stopped cow. 

Problem link: http://www.usaco.org/index.php?page=viewproblem2&cpid=1064